Tested with LIV tournament and confirmed works successfully end to end.
Improvements to console look and feel using print strings throughout.

Change made to the athlete ID following investigation of the GET response on the tournament API - confirmed same method used between PGA and LIV.

Rolled back the save to be for The Open. Might run through to fetch for those rounds to back them up.